requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
requests>=2.31.0
python-dotenv>=1.0.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Drivers para diferentes bancos de dados (descomente conforme necessário)
# psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0      # MySQL
# cx_Oracle>=8.3.0        # Oracle
